% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_definition.R
\docType{class}
\name{Box-class}
\alias{Box}
\alias{Box-class}
\title{An s4 class to represent a Box}
\description{
An s4 class to represent a Box
}
\section{Slots}{

\describe{
\item{\code{origin}}{A length-three vector}

\item{\code{length}}{A numeric}

\item{\code{height}}{A numeric}

\item{\code{width}}{A numeric}

\item{\code{weight}}{A numeric}
}}
\examples{
# create a box with size 1 x 1 x 1
b1 <- Box(length = 1, height = 1, width = 1)

}

