% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackBoxes.R
\name{CalculateFitness}
\alias{CalculateFitness}
\title{Calcuate a Fitness of Packing Solution (Percent of Wasted Space)}
\usage{
CalculateFitness(packing_solution)
}
\arguments{
\item{packing_solution}{- A list}

\item{object}{- An object of class Box or Container}
}
\value{
A number between 0 and 1
Calculate volume of an object

A numeric
}
\description{
Calcuate a Fitness of Packing Solution (Percent of Wasted Space)
}
