% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{CheckIfBoxEqualsEMS}
\alias{CheckIfBoxEqualsEMS}
\title{Check If the Box equals to the Container}
\usage{
CheckIfBoxEqualsEMS(box, ems)
}
\arguments{
\item{box}{- An object of class Box}

\item{ems}{- An object of class EMS}
}
\value{
TRUE/FALSE
}
\description{
Check If the Box equals to the Container
}
