% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_boxes.R
\name{CheckIfBoxFitsIntoEMS}
\alias{CheckIfBoxFitsIntoEMS}
\title{Check If the Box fits into the EMS}
\usage{
CheckIfBoxFitsIntoEMS(box, ems)
}
\arguments{
\item{box}{- An object of class Box}

\item{ems}{- An object of class EMS}
}
\value{
TRUE/FALSE
}
\description{
Check If the Box fits into the EMS
}
