% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{CheckIfEMSisInsideOtherEMS}
\alias{CheckIfEMSisInsideOtherEMS}
\title{Check if an EMS is totally inscribed by other EMS}
\usage{
CheckIfEMSisInsideOtherEMS(ems_to_check, ems)
}
\arguments{
\item{ems_to_check}{- An object of class EMS to check if it's in ems}

\item{ems}{- An object of class EMS}
}
\value{
TRUE if ems_to_check is inside ems otherwise FALSE
}
\description{
Check if an EMS is totally inscribed by other EMS
}
