% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_definition.R
\docType{class}
\name{Container-class}
\alias{Container}
\alias{Container-class}
\title{An s4 class to represent a Container}
\description{
An s4 class to represent a Container
}
\section{Slots}{

\describe{
\item{\code{origin}}{A length-three vector}

\item{\code{length}}{A numeric}

\item{\code{height}}{A numeric}

\item{\code{width}}{A numeric}
}}
\examples{
# create a container with size 2 x 2 x 2
c1 <- Container(length = 2, height = 2, width = 2)

}

