% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{CreateChromosome}
\alias{CreateChromosome}
\title{Create a Chromosome}
\usage{
CreateChromosome(n_boxes, n_containers)
}
\arguments{
\item{n_boxes}{- An integer}

\item{n_containers}{- An integer}
}
\value{
A list with first element BPS (Box Plasement Sequence) 
        and second element CLS (Container Loading Sequence) and
        their appropriate sequences
}
\description{
Create a Chromosome
}
