% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{CreateEMS}
\alias{CreateEMS}
\title{Create Empty Maximal Spaces}
\usage{
CreateEMS(container, box)
}
\arguments{
\item{container}{- An object of class 'Container' or 'EMS'}

\item{box}{- An object of class 'Box'}

\item{EMS}{- An instance of class EMS}
}
\value{
A list of instance of class EMS or empty list

TRUE/FALSE 
Check if EMS has nonzero parameters

TRUE if EMS is valid (all parameters are nonzero) otherwise FALSE
}
\description{
Create Empty Maximal Spaces
}
\examples{
container <- Container(width = 2, length = 4, height = 2)
box <- Box(width = 1, length = 1, height = 1, origin = c(0, 0,0))

CreateEMS(container, box)
Check if the box is otside the container

}
