% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{CustomChromosomeInitialization}
\alias{CustomChromosomeInitialization}
\title{Create 4 Initial Custom Chromosomes}
\usage{
CustomChromosomeInitialization(boxes, n_containers)
}
\arguments{
\item{boxes}{- A list of objects of class Box}
}
\value{
A list of 4 chromosomes
}
\description{
Create 4 Initial Custom Chromosomes
}
