% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{EliminateEMSList}
\alias{EliminateEMSList}
\title{Remove those EMS from list which are inside other EMS}
\usage{
EliminateEMSList(ems_list)
}
\arguments{
\item{ems_list}{- A list of objects of class EMS}
}
\value{
An updated list of EMS objects
}
\description{
Remove those EMS from list which are inside other EMS
}
