% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{InitializePopulation}
\alias{InitializePopulation}
\title{Initialize a Population}
\usage{
InitializePopulation(population_size, n_containers, boxes)
}
\arguments{
\item{population_size}{- An integer}

\item{n_containers}{- An integer}

\item{boxes}{- A list of objects of class Box}
}
\value{
A list of chromosomes
}
\description{
Initialize a Population
}
