% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{MutateChromosome}
\alias{MutateChromosome}
\title{Mutate a chromosome}
\usage{
MutateChromosome(chromosome)
}
\arguments{
\item{chromosome}{- A list with BPS and CLS}
}
\value{
A chromosome
}
\description{
Mutate a chromosome
}
