% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackBoxes.R
\name{PackBoxes}
\alias{PackBoxes}
\title{Pack the Boxes into the Containers}
\usage{
PackBoxes(boxes, containers, box_packing_sequence, container_loading_sequence)
}
\arguments{
\item{boxes}{- A list of objects of class Box}

\item{containers}{- A list of containers}

\item{box_packing_sequence}{- A vector}

\item{container_loading_sequence}{- A vector}
}
\value{
A list of Containers and corresponding Boxes placed into 
        this Containers 
Create an empty Packing Solution

A list
}
\description{
Pack the Boxes into the Containers
}
