% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{PerformCrossover}
\alias{PerformCrossover}
\title{Perform Crossover}
\usage{
PerformCrossover(mating_pool, crossover_prob)
}
\arguments{
\item{mating_pool}{- A list of chromosomes to crossover}

\item{crossover_prob}{- A numeric in [0; 1]; A probability for chromosome crossover}
}
\value{
A list of chromosomes
}
\description{
Perform Crossover
}
