% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{PerformElitism}
\alias{PerformElitism}
\title{Perform Elitism: Select the best Chromosomes within population}
\usage{
PerformElitism(fitness, elitism_size)
}
\arguments{
\item{fitness}{- A vector with chromosomes scores}

\item{elitism_size}{- An integer}
}
\value{
A vector of size 'elitism_size' with indeces of 
        best performing chromosomes
}
\description{
Perform Elitism: Select the best Chromosomes within population
}
