% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{PerformMutation}
\alias{PerformMutation}
\title{Perform Mutation}
\usage{
PerformMutation(population, mutation_prob)
}
\arguments{
\item{population}{- A list of chromosomes}

\item{mutation_prob}{- A numeric in [0; 1]; A probability for chromosome mutation}
}
\value{
A list of chromosomes
}
\description{
Perform Mutation
}
