% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_boxes.R
\name{PerformPlacementSelection}
\alias{PerformPlacementSelection}
\title{Perform Best Placement Selection}
\usage{
PerformPlacementSelection(box, ems)
}
\arguments{
\item{box}{- An object of class Box}

\item{ems}{- An object of class EMS}
}
\value{
An object of class Box
}
\description{
Perform Best Placement Selection
}
