% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_on_chromosomes.R
\name{PerformSelection}
\alias{PerformSelection}
\title{Perform Selection}
\usage{
PerformSelection(population, fitness)
}
\arguments{
\item{population}{- A list of chromosomes}

\item{fitness}{- A vector}
}
\value{
A list of chromosomes of the same size as population
}
\description{
Perform Selection
}
