% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcube.R
\name{PlotCube}
\alias{PlotCube}
\title{Plot 3D cube}
\usage{
PlotCube(object, plot_origin = TRUE, ...)
}
\arguments{
\item{object}{- An object of class 'Container', 'Box' or 'EMS'}

\item{plot_origin}{- logical, whether to plot point at the origin}
}
\description{
Plot 3D cube
}
\examples{

RGLInit(new.device = T)  # create new device with specific settings
# plot a container
container <- Container(width = 2, length = 4, height = 2)
PlotCube(container)

# plot a box
box <- Box(width = 1, length = 1, height = 1, origin = c(0, 0,0))
PlotCube(box)
}
