% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcube.R
\name{PlotPackingSolution}
\alias{PlotPackingSolution}
\title{Plot a Packing Solution}
\usage{
PlotPackingSolution(packing_solution)
}
\arguments{
\item{packing_solution}{- A list}
}
\value{
Returns as many plots of Containers with placed Boxes
        as many nonempty Containers are in the packing solution
}
\description{
Plot a Packing Solution
}
