% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{PrioritizeEMS}
\alias{PrioritizeEMS}
\title{Prioritize EMS in EMS list (sort EMS by distance to the box origin)}
\usage{
PrioritizeEMS(ems_list)
}
\arguments{
\item{ems_list}{- A list with objects of class EMS}

\item{ems}{- An object of class EMS}
}
\value{
A list with objects of class EMS
Calculate distance (Euclidean) from EMS origin to Box origin

A numeric
}
\description{
Prioritize EMS in EMS list (sort EMS by distance to the box origin)
}
