% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcube.R
\name{RGLInit}
\alias{RGLInit}
\title{A custom function to initialize RGL device}
\usage{
RGLInit(new.device = FALSE, width = 500)
}
\arguments{
\item{new.device}{- A logical value. If TRUE, creates a new device}

\item{width}{the width of the device}
}
\description{
A custom function to initialize RGL device
}
