% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMS.R
\name{UpdateEMS}
\alias{UpdateEMS}
\title{Update list of container's EMS after box is placed}
\usage{
UpdateEMS(ems_list, box)
}
\arguments{
\item{ems_list}{- A list of objects of class EMS}

\item{box}{- An object of class Box}
}
\value{
A list of objects of class EMS
}
\description{
Update list of container's EMS after box is placed
}
